refactor(runtime): merge schema package into common
