+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/common/Message.java
@@ -1,4 +1,4 @@
-package com.huawei.ascend.runtime.schema;
+package com.huawei.ascend.runtime.common;
 
 import java.util.List;
 import java.util.Map;
