+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/common/Role.java
@@ -1,4 +1,4 @@
-package com.huawei.ascend.runtime.schema;
+package com.huawei.ascend.runtime.common;
 
 /**
  * Canonical message role, shared across the agent-service modules.
