+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/session/model/Session.java
@@ -1,6 +1,6 @@
 package com.huawei.ascend.runtime.session.model;
 
-import com.huawei.ascend.runtime.schema.Message;
+import com.huawei.ascend.runtime.common.Message;
 
 import java.time.Instant;
 import java.util.List;
