+++ agent-runtime/src/test/java/com/huawei/ascend/runtime/access/protocol/a2a/egress/A2aOutputMapperTest.java
@@ -7,8 +7,8 @@ import com.huawei.ascend.runtime.access.model.NotificationType;
 import com.huawei.ascend.runtime.common.AgentResponseEvent;
 import com.huawei.ascend.runtime.common.ResponseStatus;
 import com.huawei.ascend.runtime.common.ResponseType;
-import com.huawei.ascend.runtime.schema.Message;
-import com.huawei.ascend.runtime.schema.RunStatus;
+import com.huawei.ascend.runtime.common.Message;
+import com.huawei.ascend.runtime.common.RunStatus;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
