+++ agent-runtime/src/test/java/com/huawei/ascend/runtime/bootstrap/AgentServiceEndToEndIT.java
@@ -16,8 +16,8 @@ import com.huawei.ascend.runtime.engine.spi.AgentExecutionResult;
 import com.huawei.ascend.runtime.engine.spi.AgentRuntimeHandler;
 import com.huawei.ascend.runtime.engine.spi.StreamAdapter;
 import com.huawei.ascend.runtime.queue.config.QueueAutoConfiguration;
-import com.huawei.ascend.runtime.schema.AgentRequest;
-import com.huawei.ascend.runtime.schema.Message;
+import com.huawei.ascend.runtime.common.AgentRequest;
+import com.huawei.ascend.runtime.common.Message;
 import com.huawei.ascend.runtime.session.api.SessionManager;
 import com.huawei.ascend.runtime.session.config.SessionManageConfiguration;
 import com.huawei.ascend.runtime.control.TaskControlService;
