+++ agent-runtime/src/test/java/com/huawei/ascend/runtime/control/test/TaskControlServiceWhiteboxTest.java
@@ -5,8 +5,8 @@ import com.huawei.ascend.runtime.engine.api.EnqueueEngineCancelRequest;
 import com.huawei.ascend.runtime.engine.api.EnqueueEngineExecutionRequest;
 import com.huawei.ascend.runtime.engine.api.EnqueueEngineResumeRequest;
 import com.huawei.ascend.runtime.engine.api.EnqueueEngineStatus;
-import com.huawei.ascend.runtime.schema.AgentRequest;
-import com.huawei.ascend.runtime.schema.Message;
+import com.huawei.ascend.runtime.common.AgentRequest;
+import com.huawei.ascend.runtime.common.Message;
 import com.huawei.ascend.runtime.control.Task;
 import com.huawei.ascend.runtime.control.TaskControlService;
 import com.huawei.ascend.runtime.control.TaskFailureCode;
