+++ agent-runtime/src/test/java/com/huawei/ascend/runtime/control/test/TaskflowEngineBridgeWhiteboxTest.java
@@ -10,8 +10,8 @@ import com.huawei.ascend.runtime.engine.DefaultAgentRuntimeHandlerRegistry;
 import com.huawei.ascend.runtime.engine.EngineDispatcher;
 import com.huawei.ascend.runtime.engine.support.FakeInterruptingAgentRuntimeHandler;
 import com.huawei.ascend.runtime.engine.support.RecordingAccessLayerClient;
-import com.huawei.ascend.runtime.schema.AgentRequest;
-import com.huawei.ascend.runtime.schema.Message;
+import com.huawei.ascend.runtime.common.AgentRequest;
+import com.huawei.ascend.runtime.common.Message;
 import com.huawei.ascend.runtime.control.EngineTaskControlAdapter;
 import com.huawei.ascend.runtime.control.TaskControlService;
 import com.huawei.ascend.runtime.control.TaskState;
