+++ agent-runtime/src/test/java/com/huawei/ascend/runtime/engine/adapters/openjiuwen/OpenJiuwenMessageAdapterTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.huawei.ascend.runtime.engine.handler.AgentExecutionContext;
 import com.huawei.ascend.runtime.engine.model.EngineExecutionScope;
 import com.huawei.ascend.runtime.engine.model.EngineInput;
-import com.huawei.ascend.runtime.schema.Message;
+import com.huawei.ascend.runtime.common.Message;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
